% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2norm.R
\name{estimate_initial_values_from_data}
\alias{estimate_initial_values_from_data}
\title{Get initial values from data}
\usage{
estimate_initial_values_from_data(data_list)
}
\arguments{
\item{data_list}{your inputs}
}
\value{
a named list of initial values
}
\description{
Get initial values from data
}
