% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2norm.R
\name{estimate_singleThresh_WOLS}
\alias{estimate_singleThresh_WOLS}
\title{Weighted OLS (Initial value in Single Threshold per Study MLE)}
\usage{
estimate_singleThresh_WOLS(n_i, c_i, p_i_obs)
}
\arguments{
\item{n_i}{numeric vector}

\item{c_i}{numeric vector}

\item{p_i_obs}{numeric vector}
}
\value{
A list with \code{mu}, \code{sigma}.
}
\description{
Implements the formula \eqn{ c_i = \mu + \sigma * \Phi^{-1}(1 - p_i^{obs}) }
in a weighted least-squares sense, with weights = \eqn{n_i}.
}
