% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2norm.R
\name{gaussHermite}
\alias{gaussHermite}
\title{Minimal Gauss-Hermite Quadrature}
\usage{
gaussHermite(n)
}
\arguments{
\item{n}{integer number of quadrature points}
}
\value{
list with \code{nodes} and \code{weights}
}
\description{
Returns \code{(nodes, weights)} for approximating \eqn{\int f(x) e^{-x^2} dx},
ignoring any normalizing constant. This is a simple demonstration; for serious
applications, more robust libraries or expansions might be used.
}
